==> 0_get_CoRTAD_data.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r load libraries}
library(ncdf4)
library(stringr)
```

#set path for where you keep the bleaching data
Bleaching_data_directory="C:/Users/Shannon/Desktop/Ecoregions"
#set path for the external hard drive where you have downloaded the CoRTAD data
Cortad_directory="E:/Cortad_2017"
setwd(Bleaching_data_directory)

#read in the Reef Check csv
Bleaching_Data <- read.csv(file="Reef Check Data Raw.csv", header=TRUE, sep=",")
#use just the "Bleaching (% of population)" data, not the "Bleaching (% of colony)" data
Bleaching_Data=subset(Bleaching_Data, Organism.Code=="Bleaching (% of population)")

#Note: scale of cortad is ~4 km   https://www.nodc.noaa.gov/sog/cortad/  Reef Check coordinates are to the .1 degrees,minutes,seconds... A tenth of a minute of Latitude (or tenth of a nautical mile) is about 600 feet or 185 meters (longitude at the equator)

```{r Data formatting: format lat and lon, calculate average bleaching, remove data points that have NA values for lat, lon, average bleaching, depth, or date}

#Format Longitude
for (i in 1:length(Bleaching_Data$Longitude.Degrees)){
  cell<-as.double(Bleaching_Data$Longitude.Degrees[i])+ as.double(Bleaching_Data$Longitude.Minutes[i]/60)+ as.double(Bleaching_Data$Longitude.Seconds[i]/60/60)
  if (Bleaching_Data$Longitude.Cardinal.Direction[i]=='W'){cell<-cell*-1}
  Bleaching_Data$Longitude.Degrees[i]<-cell
}

#Format Latitude
for (i in 1:length(Bleaching_Data$Latitude.Degrees)){
  cell<-as.double(Bleaching_Data$Latitude.Degrees[i])+ as.double(Bleaching_Data$Latitude.Minutes[i]/60)+ as.double(Bleaching_Data$Latitude.Seconds[i]/60/60)
  if (Bleaching_Data$Latitude.Cardinal.Direction[i]=='S'){cell<-cell*-1}
  Bleaching_Data$Latitude.Degrees[i]<-cell
}

Bleaching_Data = subset(Bleaching_Data, select = -c(Latitude.Minutes,Latitude.Seconds,Latitude.Cardinal.Direction, Longitude.Minutes, Longitude.Seconds, Longitude.Cardinal.Direction))

#Calculate average bleaching
average_bleaching<-array(data=NA,dim=length(Bleaching_Data$S1))
for (i in 1:length(average_bleaching)){
num.transects<-0
bleaching_sum<-0
if (is.na(Bleaching_Data$S1[i])==FALSE){num.transects=num.transects+1; bleaching_sum<-bleaching_sum+Bleaching_Data$S1[i]}
if (is.na(Bleaching_Data$S2[i])==FALSE){num.transects=num.transects+1; bleaching_sum<-bleaching_sum+Bleaching_Data$S2[i]}
if (is.na(Bleaching_Data$S3[i])==FALSE){num.transects=num.transects+1; bleaching_sum<-bleaching_sum+Bleaching_Data$S3[i]}
if (is.na(Bleaching_Data$S4[i])==FALSE){num.transects=num.transects+1; bleaching_sum<-bleaching_sum+Bleaching_Data$S4[i]}
average_bleaching[i]<-bleaching_sum/num.transects
}
Bleaching_Data$Average_bleaching<-as.double(average_bleaching)

#Remove NA's
Bleaching_Data=subset(Bleaching_Data, !is.na(Latitude.Degrees))
Bleaching_Data=subset(Bleaching_Data, !is.na(Longitude.Degrees))
Bleaching_Data=subset(Bleaching_Data, !is.na(Average_bleaching))
Bleaching_Data=subset(Bleaching_Data, !is.na(Depth))
Bleaching_Data=subset(Bleaching_Data, !is.na(Date))

number_of_surveys=dim(Bleaching_Data)[1]
#note: after removing rows with NA's we are left with 9215 studies
```


```{r open one NetCDF file and look at the dimensions}
setwd(Cortad_directory)
FilledSST<-nc_open("cortadv6_FilledSST.nc", write=FALSE, readunlim=TRUE, verbose=FALSE)
names(FilledSST$var) #prints a list of the variable names
#time_bounds, lat_bounds, lon_bounds, crs, land, NumberGood, AllBad, FilledSST, FilledSSTminimum, FilledSSTmaximum, FilledSSTstandardDeviation, FilledSSTmean
FilledSST_time_bounds<-ncvar_get(FilledSST, varid="time_bounds")
dim(FilledSST_time_bounds) #2 1878

FilledSST_lat_bounds<-ncvar_get(FilledSST, varid="lat_bounds")
dim(FilledSST_lat_bounds) #2 4320

FilledSST_lon_bounds<-ncvar_get(FilledSST, varid="lon_bounds")
dim(FilledSST_lon_bounds) #2 8640

FilledSST_land<-ncvar_get(FilledSST, varid="land")
#0 or 1
dim(FilledSST_land) #8640 4320
```

```{r check the grid cell size}
#checking if the grid is evenly split. it's very close. each step is between .04165649 and .04167175
difference<-array(0, dim=(length(FilledSST_lon_bounds[1,])-1))
for (i in 1:(dim(FilledSST_lon_bounds)[2]-1))
{difference[i]<-FilledSST_lon_bounds[1,(i+1)]-FilledSST_lon_bounds[1,i]}
```

```{r calculate latitude grid cell}
Bleaching_cortad_lat_cell<-array(0, dim=number_of_surveys)
lat_step<--1*(FilledSST_lat_bounds[2,dim(FilledSST_lat_bounds)[2]]-FilledSST_lat_bounds[1,1])/(dim(FilledSST_lat_bounds)[2]+1)

for (i in 1:number_of_surveys)
{
lat_grid_cell<-NA

if(is.na(Bleaching_Data$Latitude.Degrees[i]))
{lat_grid_cell<-NA}else{
n_lat_steps<-floor((FilledSST_lat_bounds[1,1]-Bleaching_Data$Latitude.Degrees[i])/lat_step+1)

  if(FilledSST_lat_bounds[1,n_lat_steps]>=Bleaching_Data$Latitude.Degrees[i])
  {
    if(FilledSST_lat_bounds[2,n_lat_steps]<=Bleaching_Data$Latitude.Degrees[i])
    {lat_grid_cell<-n_lat_steps}
    else
    {
      repeat{
        n_lat_steps=n_lat_steps+1
        if(FilledSST_lat_bounds[1,n_lat_steps]>Bleaching_Data$Latitude.Degrees[i]){
          if(FilledSST_lat_bounds[2,n_lat_steps]<=Bleaching_Data$Latitude.Degrees[i])
          {break}
        }
      }
      lat_grid_cell<-n_lat_steps
    }
    
  }
  
  if(FilledSST_lat_bounds[1,n_lat_steps]<Bleaching_Data$Latitude.Degrees[i])
  {
  repeat{
      n_lat_steps=n_lat_steps-1
      if(FilledSST_lat_bounds[1,n_lat_steps]>=Bleaching_Data$Latitude.Degrees[i])
      {
        if(FilledSST_lat_bounds[2,n_lat_steps]<=Bleaching_Data$Latitude.Degrees[i])
        {break}
      }
    }
    lat_grid_cell<-n_lat_steps
  }
}
  Bleaching_cortad_lat_cell[i]<-lat_grid_cell
}
```

```{r calculate longitude grid cell}
Bleaching_cortad_lon_cell<-array(0, dim=number_of_surveys)
lon_step<-(FilledSST_lon_bounds[1,dim(FilledSST_lon_bounds)[2]]-FilledSST_lon_bounds[1,1])/(dim(FilledSST_lon_bounds)[2]+1)

for (i in 1:length(Bleaching_Data$Longitude.Degrees))
{
lon_grid_cell<-NA

if(is.na(Bleaching_Data$Longitude.Degrees[i]))
{lon_grid_cell<-NA}else{
n_lon_steps<-floor(-1*(FilledSST_lon_bounds[1,1]-Bleaching_Data$Longitude.Degrees[i])/lon_step+1)

  if(n_lon_steps>(dim(FilledSST_lon_bounds)[2])){n_lon_steps<-(dim(FilledSST_lon_bounds)[2])}
  if(n_lon_steps<1){n_lon_steps<-1}

  if(FilledSST_lon_bounds[1,n_lon_steps]<=Bleaching_Data$Longitude.Degrees[i])
  {
    if(FilledSST_lon_bounds[2,n_lon_steps]>Bleaching_Data$Longitude.Degrees[i])
    {lon_grid_cell<-n_lon_steps}
    else
    {
      repeat{
        n_lon_steps=n_lon_steps+1
        if(n_lon_steps>(dim(FilledSST_lon_bounds)[2])){break}
        if(FilledSST_lon_bounds[1,n_lon_steps]<=Bleaching_Data$Longitude.Degrees[i]){
          if(FilledSST_lon_bounds[2,n_lon_steps]>Bleaching_Data$Longitude.Degrees[i])
          {break}
        }
      }
      lon_grid_cell<-n_lon_steps
    }
    
  }
  
  if(FilledSST_lon_bounds[1,n_lon_steps]>Bleaching_Data$Longitude.Degrees[i])
  {
  repeat{
      n_lon_steps=n_lon_steps-1
      if(n_lon_steps==0){break}
      if(FilledSST_lon_bounds[1,n_lon_steps]<=Bleaching_Data$Longitude.Degrees[i])
      {
        if(FilledSST_lon_bounds[2,n_lon_steps]>Bleaching_Data$Longitude.Degrees[i])
        {break}
      }
    }
    lon_grid_cell<-n_lon_steps
  }
}
  Bleaching_cortad_lon_cell[i]<-lon_grid_cell
}

```

```{r format Reef Check date to match the CoRTAD date format (days after 19811231)}
Bleaching_days_since_19811231<-array(0, dim=number_of_surveys)
for (i in 1:number_of_surveys)
{
date_string<-str_split(Bleaching_Data$Date[i], "-")
day_string<-date_string[[1]][1]
day_numeric<-as.numeric(day_string)
month_string<-date_string[[1]][2]
if (month_string=="Jan"){days_since_19811231_due_to_month_number<-0}
if (month_string=="Feb"){days_since_19811231_due_to_month_number<-31}
if (month_string=="Mar"){days_since_19811231_due_to_month_number<-59}
if (month_string=="Apr"){days_since_19811231_due_to_month_number<-90}
if (month_string=="May"){days_since_19811231_due_to_month_number<-120}
if (month_string=="Jun"){days_since_19811231_due_to_month_number<-151}
if (month_string=="Jul"){days_since_19811231_due_to_month_number<-181}
if (month_string=="Aug"){days_since_19811231_due_to_month_number<-212}
if (month_string=="Sep"){days_since_19811231_due_to_month_number<-243}
if (month_string=="Oct"){days_since_19811231_due_to_month_number<-273}
if (month_string=="Nov"){days_since_19811231_due_to_month_number<-304}
if (month_string=="Dec"){days_since_19811231_due_to_month_number<-334}
year_string<-date_string[[1]][3]
year_numeric<-as.numeric(year_string)
century<-1900
if(year_numeric<25) #ex. 95 would mean 1995, 25 would mean 2025
{century<-2000}
full_year<-century+year_numeric
#add in the number of leap years
#the following are leap years: 1984, 1988, 1992, 1996, 2000, 2004, 2008, 2012, 2016
leap_year_days<-0
if( (full_year>1984) | (full_year==1984 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-1}
if( (full_year>1988) | (full_year==1988 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-2}
if( (full_year>1992) | (full_year==1992 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-3}
if( (full_year>1996) | (full_year==1996 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-4}
if( (full_year>2000) | (full_year==2000 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-5}
if( (full_year>2004) | (full_year==2004 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-6}
if( (full_year>2008) | (full_year==2008 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-7}
if( (full_year>2012) | (full_year==2012 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-8}
if( (full_year>2016) | (full_year==2016 & month_string!="Jan" & month_string!="Feb"))
{leap_year_days<-9}

days_since_19811231<-((full_year-1982)*365)+days_since_19811231_due_to_month_number+day_numeric+leap_year_days

Bleaching_days_since_19811231[i]<-days_since_19811231
}
```

```{r calculate day index in the Cortad data}
Bleaching_cortad_day_index<-array(0, dim=number_of_surveys)
max_index_of_CoRTAD<-dim(FilledSST_time_bounds)[2]+1
for (i in 1:number_of_surveys)
{
Bleaching_cortad_day_index[i]<-floor((Bleaching_days_since_19811231[i]+FilledSST_time_bounds[1,1])/7)+1

if (Bleaching_cortad_day_index[i]>max_index_of_CoRTAD){Bleaching_cortad_day_index[i]<-NA}
}
```

```{r initialize some arrays to fill with the appropriate cortad data}
Bleaching_cortad_temperature<-array(NA, dim=number_of_surveys)
Bleaching_cortad_temperature_Minimum<-array(NA, dim=number_of_surveys)
Bleaching_cortad_temperature_Maximum<-array(NA, dim=number_of_surveys)
Bleaching_cortad_temperature_Mean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_temperature_standardDeviation<-array(NA, dim=number_of_surveys)

Bleaching_cortad_WindSpeed<-array(NA, dim=number_of_surveys)

Bleaching_cortad_SSTA<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_Minimum<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_Mean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_Maximum<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_StandardDeviation<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_Frequency<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_FrequencyMax<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_FrequencyMean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_FrequencyStandardDeviation<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_DHW<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_DHWMax<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_DHWMean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_SSTA_DHWStandardDeviation<-array(NA, dim=number_of_surveys)

Bleaching_cortad_TSA<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_StandardDeviation<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_Minimum<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_Maximum<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_Mean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_Frequency<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_FrequencyMax<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_FrequencyMean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_FrequencyStandardDeviation<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_DHW<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_DHWMax<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_DHWMean<-array(NA, dim=number_of_surveys)
Bleaching_cortad_TSA_DHWStandardDeviation<-array(NA, dim=number_of_surveys)

Bleaching_cortad_ClimSST<-array(NA, dim=number_of_surveys)
```

```{r open all the netcdf files you will need}
FilledSST<-nc_open("cortadv6_FilledSST.nc", write=FALSE, readunlim=TRUE, verbose=FALSE)
names(FilledSST$var) #FilledSST, FilledSSTminimum, FilledSSTmaximum, FilledSSTstandardDeviation, FilledSSTmean

SSTA<-nc_open("cortadv6_SSTA.nc", write=FALSE, readunlim=TRUE, verbose=FALSE)
names(SSTA$var) #SSTA_Minimum, SSTA_Maximum, SSTA_StandardDeviation, SSTA_Mean, SSTA_AbsoluteValueMean, SSTA_Frequency, SSTA_FrequencyMax, SSTA_FrequencyStandardDeviation, SSTA_FrequencyMean, SSTA_DHW,  SSTA_DHWMax, SSTA_DHWStandardDeviation, SSTA_DHWMean

TSA<-nc_open("cortadv6_TSA.nc", write=FALSE, readunlim=TRUE, verbose=FALSE)
names(TSA$var) #TSA, TSA_Minimum, TSA_Maximum, TSA_StandardDeviation, TSA_Mean, TSA_Frequency, TSA_FrequencyMax, TSA_FrequencyStandardDeviation, TSA_FrequencyMean, TSA_DHW, TSA_DHWMax, TSA_DHWStandardDeviation, TSA_DHWMean

WindSpeed<-nc_open("cortadv6_WindSpeed.nc", write=FALSE, readunlim=TRUE, verbose=FALSE)
names(WindSpeed$var) #wind_speed

HarmonicsClimatology<-nc_open("cortadv6_HarmonicsClimatology.nc", write=FALSE, readunlim=TRUE, verbose=FALSE)
names(HarmonicsClimatology$var) #ClimSST, AnnualAmplitudeCoefficient, AnnualPhaseCoefficient, SemiAnnualAmplitudeCoefficient, SemiAnnualPhaseCoefficient, MeanCoefficient, LSqFitFlag
```

```{r make the functions for grabbing the correct CoRTAD grid cell}
first_pass_function_Harmonics<-function(netcdf_variable_name, variable_id){
  result<-try(ncvar_get(netcdf_variable_name, varid=variable_id, start=c(Bleaching_cortad_lon_cell[i],Bleaching_cortad_lat_cell[i], (Bleaching_cortad_day_index[i]%%52+1)), count=c(1,1,1)), silent=TRUE)
  return(result)
}

first_pass_function_2d<-function(netcdf_variable_name, variable_id){
  result<-try(ncvar_get(netcdf_variable_name, varid=variable_id, start=c(Bleaching_cortad_lon_cell[i],Bleaching_cortad_lat_cell[i]), count=c(1,1)), silent=TRUE)
  return(result)
}

first_pass_function_3d<-function(netcdf_variable_name, variable_id){
  result<-try(ncvar_get(netcdf_variable_name, varid=variable_id, start=c(Bleaching_cortad_lon_cell[i],Bleaching_cortad_lat_cell[i],Bleaching_cortad_day_index[i]), count=c(1,1,1)), silent=TRUE)
  return(result)
}


second_pass_2d<-function(netcdf_variable_name, variable_id){
      expand=1
      result=NA
      repeat{
        expanded_grid<-try(ncvar_get(netcdf_variable_name, varid=variable_id, start=c((Bleaching_cortad_lon_cell[i]-expand),(Bleaching_cortad_lat_cell[i]-expand)), count=c((1+2*expand),(1+2*expand))), silent=TRUE)
        
        if(sum(is.na(expanded_grid))==((1+2*expand)*(1+2*expand)))
        {expand=expand+1
        if (expand>=3){break}
        }
        else{
        result<-mean(expanded_grid, na.rm=TRUE)
        break}
      }
    return(result)
}

second_pass_3d<-function(netcdf_variable_name, variable_id){
      expand=1
      result=NA
      repeat{
        expanded_grid<-try(ncvar_get(netcdf_variable_name, varid=variable_id, start=c((Bleaching_cortad_lon_cell[i]-expand),(Bleaching_cortad_lat_cell[i]-expand),Bleaching_cortad_day_index[i]), count=c((1+2*expand),(1+2*expand),1)), silent=TRUE)
        
        if(sum(is.na(expanded_grid))==((1+2*expand)*(1+2*expand)))
        {expand=expand+1
        if (expand>=3){break}
        }
        else{
        result<-mean(expanded_grid, na.rm=TRUE)
        break}
      }
      return(result)
}

second_pass_Harmonics<-function(netcdf_variable_name, variable_id){
      expand=1
      result=NA
      repeat{
        expanded_grid<-try(ncvar_get(netcdf_variable_name, varid=variable_id, start=c((Bleaching_cortad_lon_cell[i]-expand),(Bleaching_cortad_lat_cell[i]-expand), (Bleaching_cortad_day_index[i]%%52)+1), count=c((1+2*expand),(1+2*expand),1)), silent=TRUE)
        
        if(sum(is.na(expanded_grid))==((1+2*expand)*(1+2*expand)))
        {expand=expand+1
        if (expand>=3){break}
        }
        else{
        result<-mean(expanded_grid, na.rm=TRUE)
        break}
      }
      return(result)
}
```

```{r for each Reef Check survey grab the corresponding Cortad variables}
for (i in 1:number_of_surveys)
{
  if(!is.na(Bleaching_cortad_day_index[i]))
  {
    if(!is.na(Bleaching_cortad_lon_cell[i]))
    {
      if(!is.na(Bleaching_cortad_lat_cell[i]))
      {
        Bleaching_cortad_ClimSST[i]<-first_pass_function_Harmonics(HarmonicsClimatology, "ClimSST")
        Bleaching_cortad_temperature[i]<-first_pass_function_3d(FilledSST, "FilledSST")
        Bleaching_cortad_temperature_Minimum[i]<-first_pass_function_2d(FilledSST, "FilledSSTminimum")
        Bleaching_cortad_temperature_Maximum[i]<-first_pass_function_2d(FilledSST, "FilledSSTmaximum")
        Bleaching_cortad_temperature_Mean[i]<-first_pass_function_2d(FilledSST, "FilledSSTmean")
        Bleaching_cortad_temperature_standardDeviation[i]<-first_pass_function_2d(FilledSST, "FilledSSTstandardDeviation")
        Bleaching_cortad_WindSpeed[i]<-first_pass_function_3d(WindSpeed, "wind_speed")
        Bleaching_cortad_SSTA[i]<-first_pass_function_3d(SSTA, "SSTA")
        Bleaching_cortad_SSTA_StandardDeviation[i]<-first_pass_function_2d(SSTA, "SSTA_StandardDeviation")
        Bleaching_cortad_SSTA_Mean[i]<-first_pass_function_2d(SSTA, "SSTA_Mean")
        Bleaching_cortad_SSTA_Maximum[i]<-first_pass_function_2d(SSTA,"SSTA_Maximum")
        Bleaching_cortad_SSTA_Minimum[i]<-first_pass_function_2d(SSTA, "SSTA_Minimum")
        Bleaching_cortad_SSTA_Frequency[i]<-first_pass_function_3d(SSTA, "SSTA_Frequency")
        Bleaching_cortad_SSTA_FrequencyStandardDeviation[i]<-first_pass_function_2d(SSTA, "SSTA_FrequencyStandardDeviation")
        Bleaching_cortad_SSTA_FrequencyMax[i]<-first_pass_function_2d(SSTA, "SSTA_FrequencyMax")
        Bleaching_cortad_SSTA_FrequencyMean[i]<-first_pass_function_2d(SSTA, "SSTA_FrequencyMean")
        Bleaching_cortad_SSTA_DHW[i]<-first_pass_function_3d(SSTA, "SSTA_DHW")
        Bleaching_cortad_SSTA_DHWStandardDeviation[i]<-first_pass_function_2d(SSTA, "SSTA_DHWStandardDeviation")
        Bleaching_cortad_SSTA_DHWMax[i]<-first_pass_function_2d(SSTA, "SSTA_DHWMax")
        Bleaching_cortad_SSTA_DHWMean[i]<-first_pass_function_2d(SSTA, "SSTA_DHWMean")
        Bleaching_cortad_TSA[i]<-first_pass_function_3d(TSA, "TSA")
        Bleaching_cortad_TSA_StandardDeviation[i]<-first_pass_function_2d(TSA, "TSA_StandardDeviation")
        Bleaching_cortad_TSA_Maximum[i]<-first_pass_function_2d(TSA, "TSA_Maximum")
        Bleaching_cortad_TSA_Minimum[i]<-first_pass_function_2d(TSA, "TSA_Minimum")
        Bleaching_cortad_TSA_Mean[i]<-first_pass_function_2d(TSA, "TSA_Mean")
        Bleaching_cortad_TSA_Frequency[i]<-first_pass_function_3d(TSA, "TSA_Frequency")
        Bleaching_cortad_TSA_FrequencyStandardDeviation[i]<-first_pass_function_2d(TSA, "TSA_FrequencyStandardDeviation")
        Bleaching_cortad_TSA_FrequencyMax[i]<-first_pass_function_2d(TSA, "TSA_FrequencyMax")
        Bleaching_cortad_TSA_FrequencyMean[i]<-first_pass_function_2d(TSA, "TSA_FrequencyMean")
        Bleaching_cortad_TSA_DHW[i]<-first_pass_function_3d(TSA, "TSA_DHW")
        Bleaching_cortad_TSA_DHWStandardDeviation[i]<-first_pass_function_2d(TSA, "TSA_DHWStandardDeviation")
        Bleaching_cortad_TSA_DHWMax[i]<-first_pass_function_2d(TSA, "TSA_DHWMax")
        Bleaching_cortad_TSA_DHWMean[i]<-first_pass_function_2d(TSA, "TSA_DHWMean")
      }
    }
  }
  print(i)
}

#sum(is.na(Bleaching_cortad_temperature))  1603
#sum(is.na(Bleaching_cortad_temperature_standardDeviation)) 740
#sum(is.na(Bleaching_cortad_WindSpeed))  1
#sum(is.na(Bleaching_cortad_SSTA)) 1589
#sum(is.na(Bleaching_cortad_SSTA_StandardDeviation)) 740
#sum(is.na(Bleaching_cortad_SSTA_Mean)) 9215
#sum(is.na(Bleaching_cortad_SSTA_Maximum)) 741
#sum(is.na(Bleaching_cortad_SSTA_Frequency)) 1589
#sum(is.na(Bleaching_cortad_SSTA_FrequencyStandardDeviation)) 741
#sum(is.na(Bleaching_cortad_SSTA_FrequencyMax)) 493
#sum(is.na(Bleaching_cortad_SSTA_FrequencyMean)) 493
#sum(is.na(Bleaching_cortad_SSTA_DHW)) 1589
#sum(is.na(Bleaching_cortad_SSTA_DHWStandardDeviation)) 741
#sum(is.na(Bleaching_cortad_SSTA_DHWMax)) 741
#sum(is.na(Bleaching_cortad_SSTA_DHWMean)) 741
#sum(is.na(Bleaching_cortad_TSA)) 1589
#sum(is.na(Bleaching_cortad_TSA_StandardDeviation)) 9215
#sum(is.na(Bleaching_cortad_TSA_Maximum)) 9215
#sum(is.na(Bleaching_cortad_TSA_Mean)) 9215
#sum(is.na(Bleaching_cortad_TSA_Frequency)) 1
#sum(is.na(Bleaching_cortad_TSA_FrequencyStandardDeviation)) 9215
#sum(is.na(Bleaching_cortad_TSA_FrequencyMax)) 9215
#sum(is.na(Bleaching_cortad_TSA_FrequencyMean)) 9215
#sum(is.na(Bleaching_cortad_TSA_DHW)) 3
#sum(is.na(Bleaching_cortad_TSA_DHWStandardDeviation)) 9215
#sum(is.na(Bleaching_cortad_TSA_DHWMax)) 9215
#sum(is.na(Bleaching_cortad_TSA_DHWMean)) 9215


Bleaching_cortad_temperature<-as.numeric(Bleaching_cortad_temperature)
Bleaching_cortad_temperature_standardDeviation<-as.numeric(Bleaching_cortad_temperature_standardDeviation)
Bleaching_cortad_WindSpeed<-as.numeric(Bleaching_cortad_WindSpeed)
Bleaching_cortad_SSTA<-as.numeric(Bleaching_cortad_SSTA)
Bleaching_cortad_SSTA_StandardDeviation<-as.numeric(Bleaching_cortad_SSTA_StandardDeviation)
Bleaching_cortad_SSTA_Mean<-as.numeric(Bleaching_cortad_SSTA_Mean)
Bleaching_cortad_SSTA_Maximum<-as.numeric(Bleaching_cortad_SSTA_Maximum)
Bleaching_cortad_SSTA_Frequency<-as.numeric(Bleaching_cortad_SSTA_Frequency)
Bleaching_cortad_SSTA_FrequencyStandardDeviation<-as.numeric(Bleaching_cortad_SSTA_FrequencyStandardDeviation)
Bleaching_cortad_SSTA_FrequencyMax<-as.numeric(Bleaching_cortad_SSTA_FrequencyMax)
Bleaching_cortad_SSTA_FrequencyMean<-as.numeric(Bleaching_cortad_SSTA_FrequencyMean)
Bleaching_cortad_SSTA_DHW<-as.numeric(Bleaching_cortad_SSTA_DHW)
Bleaching_cortad_SSTA_DHWStandardDeviation<-as.numeric(Bleaching_cortad_SSTA_DHWStandardDeviation)
Bleaching_cortad_SSTA_DHWMax<-as.numeric(Bleaching_cortad_SSTA_DHWMax)
Bleaching_cortad_SSTA_DHWMean<-as.numeric(Bleaching_cortad_SSTA_DHWMean)
Bleaching_cortad_TSA<-as.numeric(Bleaching_cortad_TSA)
Bleaching_cortad_TSA_StandardDeviation<-as.numeric(Bleaching_cortad_TSA_StandardDeviation)
Bleaching_cortad_TSA_Maximum<-as.numeric(Bleaching_cortad_TSA_Maximum)
Bleaching_cortad_TSA_Mean<-as.numeric(Bleaching_cortad_TSA_Mean)
Bleaching_cortad_TSA_Frequency<-as.numeric(Bleaching_cortad_TSA_Frequency)
Bleaching_cortad_TSA_FrequencyStandardDeviation<-as.numeric(Bleaching_cortad_TSA_FrequencyStandardDeviation)
Bleaching_cortad_TSA_FrequencyMax<-as.numeric(Bleaching_cortad_TSA_FrequencyMax)
Bleaching_cortad_TSA_FrequencyMean<-as.numeric(Bleaching_cortad_TSA_FrequencyMean)
Bleaching_cortad_TSA_DHW<-as.numeric(Bleaching_cortad_TSA_DHW)
Bleaching_cortad_TSA_DHWStandardDeviation<-as.numeric(Bleaching_cortad_TSA_DHWStandardDeviation)
Bleaching_cortad_TSA_DHWMax<-as.numeric(Bleaching_cortad_TSA_DHWMax)
Bleaching_cortad_TSA_DHWMean<-as.numeric(Bleaching_cortad_TSA_DHWMean)

Bleaching_cortad_ClimSST<-as.numeric(Bleaching_cortad_ClimSST)
Bleaching_cortad_TSA_Minimum<-as.numeric(Bleaching_cortad_TSA_Minimum)
Bleaching_cortad_SSTA_Minimum<-as.numeric(Bleaching_cortad_SSTA_Minimum)
Bleaching_cortad_temperature_Minimum<-as.numeric(Bleaching_cortad_temperature_Minimum)
Bleaching_cortad_temperature_Maximum<-as.numeric(Bleaching_cortad_temperature_Maximum)
Bleaching_cortad_temperature_Mean<-as.numeric(Bleaching_cortad_temperature_Mean)
```

```{r for any grid cell that is still NA, expand to the next grid cells}
for (i in 1:number_of_surveys)
{
  if(!is.na(Bleaching_cortad_lon_cell[i]))
  {
  if(!is.na(Bleaching_cortad_lat_cell[i]))
  {
  if(!is.na(Bleaching_cortad_day_index[i]))
  {

  if(is.na(Bleaching_cortad_temperature[i]))
  {Bleaching_cortad_temperature[i]<-second_pass_3d(FilledSST, "FilledSST")}
  
  if(is.na(Bleaching_cortad_temperature_Mean[i]))
  {Bleaching_cortad_temperature_Mean[i]<-second_pass_2d(FilledSST, "FilledSSTmean")}
  
  if(is.na(Bleaching_cortad_temperature_Maximum[i]))
  {Bleaching_cortad_temperature_Maximum[i]<-second_pass_2d(FilledSST, "FilledSSTmaximum")}
  
  if(is.na(Bleaching_cortad_temperature_Minimum[i]))
  {Bleaching_cortad_temperature_Minimum[i]<-second_pass_2d(FilledSST, "FilledSSTminimum")}
  
  if(is.na(Bleaching_cortad_temperature_standardDeviation[i]))
  {Bleaching_cortad_temperature_standardDeviation[i]<-second_pass_2d(FilledSST, "FilledSSTstandardDeviation")}

  if(is.na(Bleaching_cortad_WindSpeed[i]))
  {Bleaching_cortad_WindSpeed[i]<-second_pass_3d(WindSpeed, "wind_speed")}
 
  if(is.na(Bleaching_cortad_SSTA[i]))
  {Bleaching_cortad_SSTA[i]<-second_pass_3d(SSTA, "SSTA")}

  if(is.na(Bleaching_cortad_SSTA_StandardDeviation[i]))
  {Bleaching_cortad_SSTA_StandardDeviation[i]<-second_pass_2d(SSTA, "SSTA_StandardDeviation")}

  if(is.na(Bleaching_cortad_SSTA_Mean[i]))
  {Bleaching_cortad_SSTA_Mean[i]<-second_pass_2d(SSTA, "SSTA_Mean")}

  if(is.na(Bleaching_cortad_SSTA_Maximum[i]))
  {Bleaching_cortad_SSTA_Maximum[i]<-second_pass_2d(SSTA, "SSTA_Maximum")}

  if(is.na(Bleaching_cortad_SSTA_Minimum[i]))
  {Bleaching_cortad_SSTA_Minimum[i]<-second_pass_2d(SSTA, "SSTA_Minimum")}  

  if(is.na(Bleaching_cortad_SSTA_Frequency[i]))
  {Bleaching_cortad_SSTA_Frequency[i]<-second_pass_3d(SSTA, "SSTA_Frequency")}

  if(is.na(Bleaching_cortad_SSTA_FrequencyStandardDeviation[i]))
  {Bleaching_cortad_SSTA_FrequencyStandardDeviation[i]<-second_pass_2d(SSTA, "SSTA_FrequencyStandardDeviation")}

  if(is.na(Bleaching_cortad_SSTA_FrequencyMax[i]))
  {Bleaching_cortad_SSTA_FrequencyMax[i]<-second_pass_2d(SSTA, "SSTA_FrequencyMax")}

  if(is.na(Bleaching_cortad_SSTA_FrequencyMean[i]))
  {Bleaching_cortad_SSTA_FrequencyMean[i]<-second_pass_2d(SSTA, "SSTA_FrequencyMean")}

  if(is.na(Bleaching_cortad_SSTA_DHW[i]))
  {Bleaching_cortad_SSTA_DHW[i]<-second_pass_3d(SSTA, "SSTA_DHW")}

  if(is.na(Bleaching_cortad_SSTA_DHWStandardDeviation[i]))
  {Bleaching_cortad_SSTA_DHWStandardDeviation[i]<-second_pass_2d(SSTA, "SSTA_DHWStandardDeviation")}

  if(is.na(Bleaching_cortad_SSTA_DHWMax[i]))
  {Bleaching_cortad_SSTA_DHWMax[i]<-second_pass_2d(SSTA, "SSTA_DHWMax")}

  if(is.na(Bleaching_cortad_SSTA_DHWMean[i]))
  {Bleaching_cortad_SSTA_DHWMean[i]<-second_pass_2d(SSTA, "SSTA_DHWMean")}

  if(is.na(Bleaching_cortad_TSA[i]))
  {Bleaching_cortad_TSA[i]<-second_pass_3d(TSA, "TSA")}

  if(is.na(Bleaching_cortad_TSA_StandardDeviation[i]))
  {Bleaching_cortad_TSA_StandardDeviation[i]<-second_pass_2d(TSA, "TSA_StandardDeviation")}

  if(is.na(Bleaching_cortad_TSA_Maximum[i]))
  {Bleaching_cortad_TSA_Maximum[i]<-second_pass_2d(TSA, "TSA_Maximum")}

  if(is.na(Bleaching_cortad_TSA_Mean[i]))
  {Bleaching_cortad_TSA_Mean[i]<-second_pass_2d(TSA, "TSA_Mean")}

  if(is.na(Bleaching_cortad_TSA_Minimum[i]))
  {Bleaching_cortad_TSA_Minimum[i]<-second_pass_2d(TSA, "TSA_Minimum")}

  if(is.na(Bleaching_cortad_TSA_Frequency[i]))
  {Bleaching_cortad_TSA_Frequency[i]<-second_pass_3d(TSA, "TSA_Frequency")}

  if(is.na(Bleaching_cortad_TSA_FrequencyStandardDeviation[i]))
  {Bleaching_cortad_TSA_FrequencyStandardDeviation[i]<-second_pass_2d(TSA, "TSA_FrequencyStandardDeviation")}

  if(is.na(Bleaching_cortad_TSA_FrequencyMax[i]))
  {Bleaching_cortad_TSA_FrequencyMax[i]<-second_pass_2d(TSA, "TSA_FrequencyMax")}

  if(is.na(Bleaching_cortad_TSA_FrequencyMean[i]))
  {Bleaching_cortad_TSA_FrequencyMean[i]<-second_pass_2d(TSA, "TSA_FrequencyMean")}

  if(is.na(Bleaching_cortad_TSA_DHW[i]))
  {Bleaching_cortad_TSA_DHW[i]<-second_pass_3d(TSA, "TSA_DHW")}

  if(is.na(Bleaching_cortad_TSA_DHWStandardDeviation[i]))
  {Bleaching_cortad_TSA_DHWStandardDeviation[i]<-second_pass_2d(TSA, "TSA_DHWStandardDeviation")}

  if(is.na(Bleaching_cortad_TSA_DHWMax[i]))
  {Bleaching_cortad_TSA_DHWMax[i]<-second_pass_2d(TSA, "TSA_DHWMax")}

  if(is.na(Bleaching_cortad_TSA_DHWMean[i]))
  {Bleaching_cortad_TSA_DHWMean[i]<-second_pass_2d(TSA, "TSA_DHWMean")}

  if(is.na(Bleaching_cortad_ClimSST[i]))
  {Bleaching_cortad_ClimSST[i]<-second_pass_Harmonics(HarmonicsClimatology, "ClimSST")}

  print(i)
  } #close lon
  } #close lat
  } #close day index
} #close number_of_surveys

#sum(is.na(Bleaching_cortad_temperature))  2499
#sum(is.na(Bleaching_cortad_temperature_standardDeviation)) 2497
#sum(is.na(Bleaching_cortad_WindSpeed))  2495
#sum(is.na(Bleaching_cortad_SSTA)) 2499
#sum(is.na(Bleaching_cortad_SSTA_StandardDeviation)) 2497
#sum(is.na(Bleaching_cortad_SSTA_Mean)) 2497
#sum(is.na(Bleaching_cortad_SSTA_Maximum)) 2497
#sum(is.na(Bleaching_cortad_SSTA_Frequency)) 2499
#sum(is.na(Bleaching_cortad_SSTA_FrequencyStandardDeviation)) 2497
#sum(is.na(Bleaching_cortad_SSTA_FrequencyMax)) 2497
#sum(is.na(Bleaching_cortad_SSTA_FrequencyMean)) 2497
#sum(is.na(Bleaching_cortad_SSTA_DHW)) 2499
#sum(is.na(Bleaching_cortad_SSTA_DHWStandardDeviation)) 2497
#sum(is.na(Bleaching_cortad_SSTA_DHWMax)) 2497
#sum(is.na(Bleaching_cortad_SSTA_DHWMean)) 2497
#sum(is.na(Bleaching_cortad_TSA)) 2499
#sum(is.na(Bleaching_cortad_TSA_StandardDeviation)) 2497
#sum(is.na(Bleaching_cortad_TSA_Maximum)) 2497
#sum(is.na(Bleaching_cortad_TSA_Mean)) 2497
#sum(is.na(Bleaching_cortad_TSA_Frequency)) 2499
#sum(is.na(Bleaching_cortad_TSA_FrequencyStandardDeviation)) 2497
#sum(is.na(Bleaching_cortad_TSA_FrequencyMax)) 2497
#sum(is.na(Bleaching_cortad_TSA_FrequencyMean)) 2497
#sum(is.na(Bleaching_cortad_TSA_DHW)) 2499
#sum(is.na(Bleaching_cortad_TSA_DHWStandardDeviation)) 2497
#sum(is.na(Bleaching_cortad_TSA_DHWMax)) 2497
#sum(is.na(Bleaching_cortad_TSA_DHWMean)) 2497

Bleaching_cortad_temperature<-as.numeric(Bleaching_cortad_temperature)
Bleaching_cortad_temperature_standardDeviation<-as.numeric(Bleaching_cortad_temperature_standardDeviation)
Bleaching_cortad_WindSpeed<-as.numeric(Bleaching_cortad_WindSpeed)
Bleaching_cortad_SSTA<-as.numeric(Bleaching_cortad_SSTA)
Bleaching_cortad_SSTA_StandardDeviation<-as.numeric(Bleaching_cortad_SSTA_StandardDeviation)
Bleaching_cortad_SSTA_Mean<-as.numeric(Bleaching_cortad_SSTA_Mean)
Bleaching_cortad_SSTA_Maximum<-as.numeric(Bleaching_cortad_SSTA_Maximum)
Bleaching_cortad_SSTA_Frequency<-as.numeric(Bleaching_cortad_SSTA_Frequency)
Bleaching_cortad_SSTA_FrequencyStandardDeviation<-as.numeric(Bleaching_cortad_SSTA_FrequencyStandardDeviation)
Bleaching_cortad_SSTA_FrequencyMax<-as.numeric(Bleaching_cortad_SSTA_FrequencyMax)
Bleaching_cortad_SSTA_FrequencyMean<-as.numeric(Bleaching_cortad_SSTA_FrequencyMean)
Bleaching_cortad_SSTA_DHW<-as.numeric(Bleaching_cortad_SSTA_DHW)
Bleaching_cortad_SSTA_DHWStandardDeviation<-as.numeric(Bleaching_cortad_SSTA_DHWStandardDeviation)
Bleaching_cortad_SSTA_DHWMax<-as.numeric(Bleaching_cortad_SSTA_DHWMax)
Bleaching_cortad_SSTA_DHWMean<-as.numeric(Bleaching_cortad_SSTA_DHWMean)
Bleaching_cortad_TSA<-as.numeric(Bleaching_cortad_TSA)
Bleaching_cortad_TSA_StandardDeviation<-as.numeric(Bleaching_cortad_TSA_StandardDeviation)
Bleaching_cortad_TSA_Maximum<-as.numeric(Bleaching_cortad_TSA_Maximum)
Bleaching_cortad_TSA_Mean<-as.numeric(Bleaching_cortad_TSA_Mean)
Bleaching_cortad_TSA_Frequency<-as.numeric(Bleaching_cortad_TSA_Frequency)
Bleaching_cortad_TSA_FrequencyStandardDeviation<-as.numeric(Bleaching_cortad_TSA_FrequencyStandardDeviation)
Bleaching_cortad_TSA_FrequencyMax<-as.numeric(Bleaching_cortad_TSA_FrequencyMax)
Bleaching_cortad_TSA_FrequencyMean<-as.numeric(Bleaching_cortad_TSA_FrequencyMean)
Bleaching_cortad_TSA_DHW<-as.numeric(Bleaching_cortad_TSA_DHW)
Bleaching_cortad_TSA_DHWStandardDeviation<-as.numeric(Bleaching_cortad_TSA_DHWStandardDeviation)
Bleaching_cortad_TSA_DHWMax<-as.numeric(Bleaching_cortad_TSA_DHWMax)
Bleaching_cortad_TSA_DHWMean<-as.numeric(Bleaching_cortad_TSA_DHWMean)

Bleaching_cortad_ClimSST<-as.numeric(Bleaching_cortad_ClimSST)
Bleaching_cortad_TSA_Minimum<-as.numeric(Bleaching_cortad_TSA_Minimum)
Bleaching_cortad_SSTA_Minimum<-as.numeric(Bleaching_cortad_SSTA_Minimum)
Bleaching_cortad_temperature_Minimum<-as.numeric(Bleaching_cortad_temperature_Minimum)
Bleaching_cortad_temperature_Maximum<-as.numeric(Bleaching_cortad_temperature_Maximum)
Bleaching_cortad_temperature_Mean<-as.numeric(Bleaching_cortad_temperature_Mean)
```

```{r format the Reef Check Bleaching data and Cortad variables and write a csv}
setwd(Bleaching_data_directory)

Bleaching_Data_with_cortad_variables<-cbind(Bleaching_Data, Bleaching_cortad_ClimSST, Bleaching_cortad_temperature, Bleaching_cortad_temperature_Mean, Bleaching_cortad_temperature_Minimum, Bleaching_cortad_temperature_Maximum, Bleaching_cortad_temperature_standardDeviation, Bleaching_cortad_WindSpeed, Bleaching_cortad_SSTA, Bleaching_cortad_SSTA_StandardDeviation, Bleaching_cortad_SSTA_Mean, Bleaching_cortad_SSTA_Minimum, Bleaching_cortad_SSTA_Maximum, Bleaching_cortad_SSTA_Frequency, Bleaching_cortad_SSTA_FrequencyStandardDeviation, Bleaching_cortad_SSTA_FrequencyMax, Bleaching_cortad_SSTA_FrequencyMean, Bleaching_cortad_SSTA_DHW, Bleaching_cortad_SSTA_DHWStandardDeviation, Bleaching_cortad_SSTA_DHWMax, Bleaching_cortad_SSTA_DHWMean, Bleaching_cortad_TSA, Bleaching_cortad_TSA_StandardDeviation, Bleaching_cortad_TSA_Minimum, Bleaching_cortad_TSA_Maximum, Bleaching_cortad_TSA_Mean, Bleaching_cortad_TSA_Frequency, Bleaching_cortad_TSA_FrequencyStandardDeviation, Bleaching_cortad_TSA_FrequencyMax, Bleaching_cortad_TSA_FrequencyMean, Bleaching_cortad_TSA_DHW, Bleaching_cortad_TSA_DHWStandardDeviation, Bleaching_cortad_TSA_DHWMax, Bleaching_cortad_TSA_DHWMean)

number_of_columns<-dim(Bleaching_Data)[2]
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+1]<-"ClimSST"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+2]<-"Temperature_Kelvin"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+3]<-"Temperature_Mean"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+4]<-"Temperature_Minimum"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+5]<-"Temperature_Maximum"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+6]<-"Temperature_Kelvin_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+7]<-"Windspeed"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+8]<-"SSTA"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+9]<-"SSTA_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+10]<-"SSTA_Mean"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+11]<-"SSTA_Minimum"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+12]<-"SSTA_Maximum"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+13]<-"SSTA_Frequency"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+14]<-"SSTA_Frequency_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+15]<-"SSTA_FrequencyMax"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+16]<-"SSTA_FrequencyMean"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+17]<-"SSTA_DHW"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+18]<-"SSTA_DHW_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+19]<-"SSTA_DHWMax"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+20]<-"SSTA_DHWMean"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+21]<-"TSA"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+22]<-"TSA_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+23]<-"TSA_Minimum"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+24]<-"TSA_Maximum"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+25]<-"TSA_Mean"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+26]<-"TSA_Frequency"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+27]<-"TSA_Frequency_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+28]<-"TSA_FrequencyMax"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+29]<-"TSA_FrequencyMean"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+30]<-"TSA_DHW"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+31]<-"TSA_DHW_Standard_Deviation"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+32]<-"TSA_DHWMax"
colnames(Bleaching_Data_with_cortad_variables)[number_of_columns+33]<-"TSA_DHWMean"

write.csv(Bleaching_Data_with_cortad_variables, file = "Reef_Check_with_cortad_variables.csv")
```

```{r close the netcdf files}
setwd(Cortad_directory)
nc_close(FilledSST)
nc_close(HarmonicsClimatology)
nc_close(SSTA)
nc_close(TSA)
nc_close(WindSpeed)
```
